+++ src/test/java/org/fxmisc/flowless/CellCreationAndLayoutEfficiencyTest.java
@@ -36,7 +36,7 @@ public class CellCreationAndLayoutEfficiencyTest {
         for(int i = 0; i < 20; ++i) {
             items.addAll("red", "green", "blue", "purple");
         }
-        VirtualFlow<String, ?> flow = VirtualFlow.createVertical(
+        VirtualizedScrollPane<String, ?> flow = VirtualizedScrollPane.createVertical(
                 items,
                 color -> {
                     cellCreations.inc();
